support HTML parsing and validations on the BST
The vscode HTML service is cool and what not but its not really all that powerful. The original approach was the leverage the HTML Language service for features but this was just overworked.

This commit provides the initial parse approach for HTML syntax and in addition will provide validations and lint capabilities. Liquid within HTML is common and while the vscode-html-language-service is a one-size fits all module/parser (including within Liquid) it is far from strict in the sense that one can express incorrect syntax, eg: `<44-# %:dd>` would be considered valid within the html language service, which its fucking not.

The HTML parsing aspect will play a strong role, it will allow all capabilities so neatly fit together without having to reach for the service and run multiple text document parses for basic features.

This approach will also allow Liquify to tightly couple with HTML and act as 1 language.

fuck my life.